restructure lazy logicalplan module
