+++ polars/polars-lazy/src/logical_plan/lit.rs
@@ -1,0 +1,151 @@
+use crate::prelude::*;
+#[cfg(feature = "temporal")]
+use polars_core::export::chrono::{NaiveDate, NaiveDateTime};
+use polars_core::prelude::*;
+
+#[derive(Clone, Debug, PartialEq)]
+pub enum LiteralValue {
+    Null,
+    /// A binary true or false.
+    Boolean(bool),
+    /// A UTF8 encoded string type.
+    Utf8(String),
+    /// An unsigned 8-bit integer number.
+    #[cfg(feature = "dtype-u8")]
+    UInt8(u8),
+    /// An unsigned 16-bit integer number.
+    #[cfg(feature = "dtype-u16")]
+    UInt16(u16),
+    /// An unsigned 32-bit integer number.
+    UInt32(u32),
+    /// An unsigned 64-bit integer number.
+    UInt64(u64),
+    /// An 8-bit integer number.
+    #[cfg(feature = "dtype-i8")]
+    Int8(i8),
+    /// A 16-bit integer number.
+    #[cfg(feature = "dtype-i16")]
+    Int16(i16),
+    /// A 32-bit integer number.
+    Int32(i32),
+    /// A 64-bit integer number.
+    Int64(i64),
+    /// A 32-bit floating point number.
+    Float32(f32),
+    /// A 64-bit floating point number.
+    Float64(f64),
+    Range {
+        low: i64,
+        high: i64,
+        data_type: DataType,
+    },
+    #[cfg(all(feature = "temporal", feature = "dtype-datetime"))]
+    DateTime(NaiveDateTime),
+    Series(NoEq<Series>),
+}
+
+impl LiteralValue {
+    /// Getter for the `DataType` of the value
+    pub fn get_datatype(&self) -> DataType {
+        match self {
+            LiteralValue::Boolean(_) => DataType::Boolean,
+            #[cfg(feature = "dtype-u8")]
+            LiteralValue::UInt8(_) => DataType::UInt8,
+            #[cfg(feature = "dtype-u16")]
+            LiteralValue::UInt16(_) => DataType::UInt16,
+            LiteralValue::UInt32(_) => DataType::UInt32,
+            LiteralValue::UInt64(_) => DataType::UInt64,
+            #[cfg(feature = "dtype-i8")]
+            LiteralValue::Int8(_) => DataType::Int8,
+            #[cfg(feature = "dtype-i16")]
+            LiteralValue::Int16(_) => DataType::Int16,
+            LiteralValue::Int32(_) => DataType::Int32,
+            LiteralValue::Int64(_) => DataType::Int64,
+            LiteralValue::Float32(_) => DataType::Float32,
+            LiteralValue::Float64(_) => DataType::Float64,
+            LiteralValue::Utf8(_) => DataType::Utf8,
+            LiteralValue::Range { data_type, .. } => data_type.clone(),
+            #[cfg(all(feature = "temporal", feature = "dtype-datetime"))]
+            LiteralValue::DateTime(_) => DataType::Datetime,
+            LiteralValue::Series(s) => s.dtype().clone(),
+            LiteralValue::Null => DataType::Null,
+        }
+    }
+}
+
+pub trait Literal {
+    /// [Literal](Expr::Literal) expression.
+    fn lit(self) -> Expr;
+}
+
+impl Literal for String {
+    fn lit(self) -> Expr {
+        Expr::Literal(LiteralValue::Utf8(self))
+    }
+}
+
+impl<'a> Literal for &'a str {
+    fn lit(self) -> Expr {
+        Expr::Literal(LiteralValue::Utf8(self.to_owned()))
+    }
+}
+
+macro_rules! make_literal {
+    ($TYPE:ty, $SCALAR:ident) => {
+        impl Literal for $TYPE {
+            fn lit(self) -> Expr {
+                Expr::Literal(LiteralValue::$SCALAR(self))
+            }
+        }
+    };
+}
+
+make_literal!(bool, Boolean);
+make_literal!(f32, Float32);
+make_literal!(f64, Float64);
+#[cfg(feature = "dtype-i8")]
+make_literal!(i8, Int8);
+#[cfg(feature = "dtype-i16")]
+make_literal!(i16, Int16);
+make_literal!(i32, Int32);
+make_literal!(i64, Int64);
+#[cfg(feature = "dtype-u8")]
+make_literal!(u8, UInt8);
+#[cfg(feature = "dtype-u16")]
+make_literal!(u16, UInt16);
+make_literal!(u32, UInt32);
+make_literal!(u64, UInt64);
+
+/// The literal Null
+pub struct Null {}
+
+impl Literal for Null {
+    fn lit(self) -> Expr {
+        Expr::Literal(LiteralValue::Null)
+    }
+}
+
+#[cfg(all(feature = "temporal", feature = "dtype-datetime"))]
+impl Literal for NaiveDateTime {
+    fn lit(self) -> Expr {
+        Expr::Literal(LiteralValue::DateTime(self))
+    }
+}
+
+#[cfg(all(feature = "temporal", feature = "dtype-datetime"))]
+impl Literal for NaiveDate {
+    fn lit(self) -> Expr {
+        Expr::Literal(LiteralValue::DateTime(self.and_hms(0, 0, 0)))
+    }
+}
+
+impl Literal for Series {
+    fn lit(self) -> Expr {
+        Expr::Literal(LiteralValue::Series(NoEq::new(self)))
+    }
+}
+
+/// Create a Literal Expression from `L`
+pub fn lit<L: Literal>(t: L) -> Expr {
+    t.lit()
+}
