+++ polars/polars-lazy/src/prelude.rs
@@ -5,7 +5,7 @@ pub use crate::{
     frame::*,
     logical_plan::{
         optimizer::{type_coercion::TypeCoercionRule, Optimize, *},
-        DataFrameUdf, LiteralValue, LogicalPlan, LogicalPlanBuilder,
+        *,
     },
     physical_plan::{expressions::*, planner::DefaultPlanner, Executor, PhysicalPlanner},
 };
